backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==25.3.0
packaging==26.2
psycopg2==2.9.11
python-dotenv==1.2.2
python-magic==0.4.27
sqlparse==0.5.5
tzdata==2025.3
